+++ tests/testthat/test-categories.R
@@ -12,6 +12,13 @@ with_mock_HTTP({
         expect_length(cats, 3)
     })
 
+    test_that("Categories print method", {
+        expect_output(cats[[1]],
+            get_output(data.frame(id=1, name="Male", value=1, missing=FALSE)))
+        expect_output(cats,
+            get_output(data.frame(id=c(1, 2, -1), name=c("Male", "Female", "No Data"), value=c(1, 2, NA), missing=c(FALSE, FALSE, TRUE))))
+    })
+
     test_that("Categories validation", {
         expect_error(Categories(
             list(id=-1L, name="B", numeric_value=1L, missing=FALSE),