+++ NEWS.md
@@ -1,5 +1,6 @@
 ### crunch 1.13.9 (under development)
 * Prevent accidentally deleting all rows with a bad expression via `dropRows`
+* Print methods for Category/ies now show category ids
 
 ### crunch 1.13.8
 * `crunchBox` to make a public, embeddable analysis widget